requirements.txt
# Trader v2.2 依赖
# pip install -r requirements.txt --break-system-packages

# 核心
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
PyYAML>=6.0
requests>=2.28.0

# 数据库
# sqlite3 内置

# ML
scikit-learn>=1.2.0
lightgbm>=4.0.0          # Phase 2: 替代 sklearn GBM (可选, 自动降级)

# WebSocket
websockets>=12.0

# 绘图 (可选)
# matplotlib>=3.7.0

# 开发
pytest>=7.0.0
